Math UPDATE
Update some features in the math section to sort traces based on
properties. Also added much more support for Traces, called TraceSection
and their plural form TraceSections. The first models a group of traces
that become some sort of a sub-section of the project. The latter can be
seen as a group of sections, or a whole project. The supported methods
currently focus on sorting properly. IT IS NOT YET FINISHED!!!!. Many
errors still exist in the code, they need to be fixed first.