Changed the location where the intersection between line and cicle is calculated
Also, the linetrace calculated has overestimated the distance to the
trace by a factor of 2, to reduce the amount of missed thingies. However
in cases of circles this could lead to a sever overestimate. Therefor we
calculate the perpendicular distance between the robot arm and the trace
before overestimating the distance to the trace.