2 things are changed
fixed the shit were the getMotor() stuff was working as a setter and
geter. BaseJoint has a std::shared_ptr motor which points to the motor
on the heap. Each joint implementation ensures a non-null shared
pointers is given to the constructor of base joint which copies the
shared pointer to its member. In this manner the setter and getter are
properly sepperated. Changes in the speed controllers are made to
accomidate this change. Equally for the IO

Secondly, getJointStatus(PinState) is implemented. This function will
return if the joint is enabled, if so, it will update the direction
which is traveling. To do this, a fellow method for the motor is
implemented.Same for the fully implemented StepperDriver