fixed calibration count bug and enable state in gui
Fixed a bug in the director where the counting did not apear to be
correct. WHen both the end stop calibrations are in place, it does not
mater if the point calibration is set to be used because it becomes
redundant. So it should not be counted as 1.
Updated the gui that, when the end stop position calibrations is not
selected, the movement variant gets disabled and removed from the
director.